+++ packages/schema/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @powersync/service-schema
 
+## 1.15.8
+
 ## 1.15.7
 
 ## 1.15.6